// src/lib/models/IClip.ts
export interface IClip {
    broadcaster_id: string;
    broadcaster_name: string;
    created_at: string;
    creator_id: string;
    creator_name: string;
    duration: number;
    embed_url: string;
    game_id: null | string;
    id: string;
    is_featured: boolean;
    language: string;
    thumbnail_url: string;
    title: string;
    url: string;
    video_id: null | string;
    view_count: number;
    vod_offset: null | number;
}